+++ CHANGELOG.md
@@ -8,19 +8,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
-- Added ignoring notifications for currently displayed chat [#799](https://github.com/parres-hq/whitenoise_flutter/issues/799)
+
 - Added automatic notification clearing when entering chat - notifications for messages and invites now automatically disappear when viewing the related chat [#803](https://github.com/parres-hq/whitenoise_flutter/issues/803)
 - Added unread counts for DMs and groups [[#806](https://github.com/parres-hq/whitenoise_flutter/issues/806)]
-- Added zoom for images in chat [#834](https://github.com/parres-hq/whitenoise_flutter/pull/834)
-- Added media icon in chat list last message preview [#841](https://github.com/parres-hq/whitenoise_flutter/pull/841)
+- Added deep linking for notifications - tapping a message or invite notification now routes directly to the correct chat or invite screen [#781](https://github.com/parres-hq/whitenoise_flutter/issues/781)
 
 ### Changed
+
 ### Deprecated
+
 ### Removed
+
 ### Fixed
+
 - Fixed keypackage operation buttons showing loading state incorrectly
 - Fixed developer settings content going under three-button bottom nav on GrapheneOS and some Android devices
 - Fixed messages disappearing for a while after being sent [#828](https://github.com/parres-hq/whitenoise_flutter/pull/828)
+
 ### Security
 
 ## [0.2.0] - 2025-11-07